web_api/requirements.txt
fastapi
pydantic
uvicorn
google-cloud-iap
